74 pulizia codice e verifica commenti

+++ MegaDex/src/main/java/com/restart/controller/DeckController.java
@@ -2,13 +2,12 @@ package com.restart.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+
 
 import java.util.List;
 import java.util.Map;
@@ -16,13 +15,11 @@ import java.util.Map;
 import com.restart.entity.DeckPass;
 import com.restart.entity.Deck;
 import com.restart.entity.User;
-import com.restart.entity.Card;
 import com.restart.entity.Slot;
 import com.restart.service.DeckServiceImpl;
 import com.restart.service.UserServiceImpl;
 import com.restart.service.CardServiceImpl;
 import com.restart.service.SlotServiceImpl;
-import com.restart.service.UserService;
 
 @RestController
 @RequestMapping("/api")
@@ -33,7 +30,7 @@ public class DeckController {
 	@Autowired
 	private DeckServiceImpl deckService;
 	@Autowired
-  private SlotServiceImpl slotService;
+  	private SlotServiceImpl slotService;
   @Autowired
   private CardServiceImpl cardService;
   
@@ -54,43 +51,54 @@ public class DeckController {
   
 	@PostMapping("/auth/addDeck")
 	public ResponseEntity<Deck> addDeck(@RequestBody Deck deck){
-		
-		//Gets the user from the database through the id sent with the request
-		/*
-		User user = userService.findUserById(deck.getUser().getId())
-				.orElseThrow(() -> new RuntimeException("User not found with ID: " + deck.getUser().getId()));
-				*/
+		//Recupera utente autenticato
 		User user = userService.getAuthenticatedUser();
-		
-
-
-		//Creates the new deck and sets the user associated to the deck
 		deck.setUser(user);
-		
-		//Saves the new deck
-		Deck newDeck = deckService.saveDeck(deck);
-		return ResponseEntity.ok(newDeck);
+
+		//Associa la lista di slot se esistente
+		if(deckService.getDeckById(deck.getId()).isPresent()){
+				if(deckService.getDeckById(deck.getId()).get().getUser() != user){
+					throw new RuntimeException("Id Deck not matching its owner");
+				}
+            deck.setSlots(deckService.getDeckById(deck.getId()).get().getSlots());
+        }
+
+        //Salva il deck
+        Deck newDeck = deckService.saveDeck(deck);
+        return ResponseEntity.ok(newDeck);
 	}
 	
-
+//Rimuove un deck
 	@PostMapping("/deb/removeDeck")
 	public ResponseEntity<String> removeDeck(@RequestBody Deck deck){
-		
-		//Removes the deck
+		//Recupera utente autenticato
+		User user = userService.getAuthenticatedUser();
+		deck.setUser(user);
+
+		//Controlla che il proprietario sia quello giusto
+		if(deckService.getDeckById(deck.getId()).isPresent()
+				&& deckService.getDeckById(deck.getId()).get().getUser() != user){
+					throw new RuntimeException("Id Deck not matching its owner");
+		}
+
+		//Rimuove il deck
 		try {
 			deckService.removeDeck(deck);
-			return ResponseEntity.ok("Deck eliminato con successo");
+			return ResponseEntity.ok("Deck erased successfully");
 		} catch (Exception e) {
 			return ResponseEntity.noContent().build();
 		}
 	}
-  
+
+	//Valida il set di un deck secondo il regolamento
 	@PostMapping("/deb/validateDeck")
     public DeckPass validateDeck(@RequestBody List<Slot> deck) {
-      for(Slot slot : deck) {
+      //Associa le carte al deck
+		for(Slot slot : deck) {
             slot.setCard(cardService.getCardById(slot.getId().getIdCard())
                     .orElseThrow(() -> new RuntimeException("Card not found with ID: " + slot.getId().getIdCard())));
         }  
-      return slotService.validateSlots(deck);
+      //Esegue la validazione del deck e ritorna un DeckPass
+		return slotService.validateSlots(deck);
     }
 }
